requirements.txt
gym==0.17.3
tensorflow-gpu
matplotlib
numpy